requirements.txt
streamlit
opencv-python
numpy
Pillow
pandas
plotly
scikit-learn
